AF-4:  2. shared -> services        3. shared -> models        + Reworked structure of the server        + Created models of entities: user, author, comments
