+++ afisha-client/src/app/app.module.ts
@@ -11,7 +11,7 @@ import {appRoutes} from "./app.routes";
 import {NavbarComponent} from './navbar/navbar.component';
 import {EventsListComponent} from './events/events-list.component';
 import {EventDetailsComponent} from './events/event-details.component';
-import {EventService} from "./events/services/event.service";
+import {EventService} from "./shared/services/event.service";
 import {EventComponent} from './events/event.component';
 import {GuardEventRouteService} from "./shared/guards/guard-event-route.service";
 import {SearchComponent} from './search/search.component';
