--- afisha-client/src/app/events/services/event.model.ts
@@ -1,13 +1,0 @@
-export interface IEvent {
-  id: number,
-  title: string,
-  shortDescription: string,
-  description: string,
-  createdAt: string,
-  updatedAt: string,
-  date: string,
-  imageUrl: string,
-  location?: {
-    address: string
-  }
-}
