+++ afisha-client/src/app/shared/models/event.model.ts
@@ -1,0 +1,13 @@
+export interface IEvent {
+  id: number,
+  title: string,
+  shortDescription: string,
+  description: string,
+  createdAt: string,
+  updatedAt: string,
+  date: string,
+  imageUrl: string,
+  location?: {
+    address: string
+  }
+}
