--- afisha-server/controllers/eventController.js
@@ -1,5 +1,0 @@
-const eventController = {
-    //@TODO methods to work with Events from DB
-};
-
-module.exports = eventController;
