--- afisha-server/gulpfile.js
@@ -1,16 +1,0 @@
-const gulp = require('gulp'),
-    nodemon = require('gulp-nodemon');
-
-gulp.task('default', () => {
-    nodemon({
-        script: 'index.js',
-        ext: 'js',
-        env: {
-            PORT: 8000
-        },
-        ignore: ['./node_modules/**']
-    })
-    .on('restart', () => {
-        console.log("Restarting");
-    });
-});
