+++ afisha-server/index.js
@@ -1,23 +1,10 @@
-const express = require('express'),
-      mongoose = require('mongoose'),
-      bodyParser = require('body-parser'),
-      indexRoutes = require('./routes/indexRoutes');
+//intro point for our server
+const config = require('./server/config/config');
+const app = require('./server/server');
+const mongoose = require('mongoose');
 
-const db = mongoose.connect('mongodb://localhost/eventAPI');
-const app = express();
-mongoose.Promise = require('bluebird');
+mongoose.connect(config.dataBaseUrl);
 
-let port = process.env.PORT || 3000;
-
-app.use(bodyParser.urlencoded({extended: true}));
-app.use(bodyParser.json());
-
-app.use('/api', indexRoutes);
-
-app.get('/*', (req, res) => {
-    res.redirect('/api');
-});
-
-app.listen(port, () => {
-    console.log("Running on http://localhost:" + port);
+app.listen(config.port, () => {
+    console.log('Running on http://localhost:' + config.port);
 });
