+++ afisha-server/package.json
@@ -19,7 +19,9 @@
   "dependencies": {
     "bluebird": "^3.5.1",
     "body-parser": "^1.18.2",
+    "colors": "^1.2.1",
     "express": "^4.16.3",
+    "lodash": "^4.17.10",
     "mongoose": "^5.0.14",
     "npm-scripts": "^1.4.0"
   }
