+++ afisha-server/server/api/controllers/eventController.js
@@ -1,0 +1,5 @@
+const eventController = {
+    //@TODO methods to work with Events from DB
+};
+
+module.exports = eventController;
