+++ afisha-server/server/config/development.js
@@ -1,0 +1,4 @@
+module.exports = {
+    //enable logging for development
+    logging: true
+};
