+++ afisha-server/server/config/production.js
@@ -1,0 +1,4 @@
+module.exports = {
+    //enable logging for development
+    logging: false
+};
