+++ afisha-server/server/config/testing.js
@@ -1,0 +1,4 @@
+module.exports = {
+    //enable logging for development
+    logging: true
+};
