+++ afisha-server/server/middleware/appMiddleware.js
@@ -1,0 +1,7 @@
+const bodyParser = require('body-parser');
+//setup global middleware here
+
+module.exports = (app) => {
+    app.use(bodyParser.urlencoded({extended: true}));
+    app.use(bodyParser.json());
+};
