+++ afisha-server/server/middleware/errorMiddleware.js
@@ -1,0 +1,6 @@
+module.exports = () => {
+    return (err, req, res, next) => {
+        console.log(err.message);
+        res.status(500);
+    };
+};
